+++ packages/docs/.storybook/main.ts
@@ -1,31 +1,32 @@
 import type { StorybookConfig } from '@storybook/html-vite';
 
 const config: StorybookConfig = {
+  addons: [
+    "@storybook/addon-links",
+    "@storybook/addon-essentials",
+    "@storybook/addon-interactions",
+    "@storybook/addon-a11y",
+    '@geometricpanda/storybook-addon-badges',
+  ],
+  core: {
+    disableTelemetry: true,
+  },
+  docs: {
+    autodocs: "tag",
+  },
+  framework: {
+    name: "@storybook/html-vite",
+    options: {},
+  },
+  staticDirs: [
+    '../public',
+  ],
   stories: [
     "../stories/**/*.mdx",
     "../stories/**/*.stories.@(js|jsx|mjs|ts|tsx)",
     "../src/**/*.mdx",
     "../src/**/*.stories.@(js|jsx|mjs|ts|tsx)",
   ],
-  staticDirs: [
-    '../public',
-  ],
-  addons: [
-    "@storybook/addon-links",
-    "@storybook/addon-essentials",
-    "@storybook/addon-interactions",
-    "@storybook/addon-a11y",
-  ],
-  framework: {
-    name: "@storybook/html-vite",
-    options: {},
-  },
-  docs: {
-    autodocs: "tag",
-  },
-  core: {
-    disableTelemetry: true,
-  },
 };
 
 export default config;